Fix analyser
